+++ Unite.Genome.Annotations/Services/Vep/SvsAnnotationService.cs
@@ -26,7 +26,7 @@ public class SvsAnnotationService
     }
 
 
-    public EffectsDataModel[] Annotate(long[] variantIds)
+    public EffectsDataModel[] Annotate(int[] variantIds)
     {
         var variants = LoadVariants(variantIds);
 
@@ -38,7 +38,7 @@ public class SvsAnnotationService
     }
 
 
-    private Variant[] LoadVariants(long[] variantIds)
+    private Variant[] LoadVariants(int[] variantIds)
     {
         using var dbContext = _dbContextFactory.CreateDbContext();
